pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-composition-patterns"
version = "0.1.0"
description = "A practical demonstration of LLM workflow patterns and agent architectures inspired by Anthropic's 'Building Effective Agents' article"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.49.0",
    "fireworks-ai>=0.15.12",
    "groq>=0.18.0",
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.36",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]
